Switch backend API to v2 and drop v1
